Added redirects for Passenger
When serving this application out through Passenger, certain actions
would lead to a HTTP 500 error. After inspecting the Rack::Request.env
attribute, it appears that request_path is missing and/or not set when
using Passenger. And since this value is used to redirect the user in
certain cases, Rack was throwing an error because it was nil. The fix
was to statically set the redirect variable for certain actions.