Changed default application port to 3000
This change updates the default port for the thin web server from 7143 to 3000 across config files and code examples.

The original intention was to avoid clashing with other services running locally. With the advent of Container Technology, that's less common than it used to be.

The port was changed to align with common conventions for development webservers. Specifying a well-known port will make the application easier for new developers to run locally, without needing to manually configure the port number.