Update references to stale env vars
Standardize naming of environment variables used to configure the app across Dockerfile and usage instructions. References to ONETIMESECRET_ were replaced with the generic REDIS_ and SECRET to avoid overspecifying the app name in the configs. This makes the environment configuration more consistent and reusable for deploying variations of the web app.

The diff updates:

- References to ONETIMESECRET_REDIS_URL and ONETIMESECRET_SECRET in the Dockerfile
- References to ONETIMESECRET_REDIS_URL, ONETIMESECRET_SSL, ONETIMESECRET_HOST, and ONETIMESECRET_SECRET in the usage instructions

To use the more generic names REDIS_URL, REDIS, HOST, and SECRET, which is clearer since these vars configure common services like Redis rather than being specific to OneTimeSecret. This avoid too tightly coupling the environment configuration to a particular app implementation.