[#321] Introduce path configuration option
Allow the directory path for onetime configuration files to be
configured directly. Previously this path was selected from a hardcoded
list, but the tryouts needed more flexibility in where these files are
located. The new attr_writer permits setting :path for Onetine::Config
to customize its file search locations.

NOTE: When the config file is specified, the locale directory is
expected to have the same parent dir.