[#321] Update Docker build to publish images to Docker Hub
This change adds configuration to publish the Docker image to Docker Hub
(as well as GitHub Container Registry). Publishing to both registries
allows for wider distribution and discoverability of the image.

The Dockerfile and build workflows were modified to:

- Define a new `DOCKERHUB_IMAGE_NAME` environment variable equal to the
repository name
- Update the Docker metadata action to target both the GitHub and Docker
Hub image paths

This ensures the image will be built and tagged for both target
registries on each release.