[#321] Reconcile local and CI test configs
Redis specifically. The crucial part is understanding how to connect to
the redis service container during GitHub Action runs. There's
supposedly a hostname to connect to but what worked out of the box is
port mapping. It works the same as docker, where the mapped port becomes
available via 127.0.0.1 on the host.

Details here:
https://docs.github.com/en/actions/using-containerized-services/creating-redis-service-containers#configuring-the-runner-job