[#392] Add truemail dependency
Rationale: truemail is a gem that goes above and beyond when it comes to
validating email addresses. It checks the domain, the MX records, and
the SMTP server to ensure that the email address is valid. We will use
it to ensure that the email addresses being used by the application are
valid. This will help us to avoid sending emails to invalid email
addresses and improve our mail deliverability.