[#392] Improve start-time nomenclature
Changed start methods to clarify boot process:

- Renamed Onetime.load! to Onetime.boot! to better reflect its purpose
of initializing the application
- Updated references to the boot method throughout the codebase
- Added documentation to Onetime module and boot method describing the
boot process

This simplifies initialization and improves grokability by using more
precise terminology. The name "load" was somewhat ambiguous, while
"boot" more clearly conveys that the application is being started up and
prepared to serve requests. "Boot" also implies there's an order of
operations which is important to understand when working with the
application.

Notes:

- No functional changes, just naming for clarity
- More boot process documentation to be added