[#383] Add package rule to mortalize 'encryptor'
Ruby's encryptor is a core package that we need to be very specific
about how we handle. This codebase is pinned to encryptor 1.x and we
need to ensure that we don't accidentally (or automatically upgrade).

When we do upgrade, we will skip encryptor 2 based on their
recommendation to avoid it and go straight to encryptor 3.

This change will allow us to close the immortal, auto-update PRs from
renovate.