[#454] Fix email subject to reference customer ID
This change updates the email subject and body text to reference the
customer ID instead of sender name to avoid confusion over who initiated
the secret. The customer ID provides a clearer indicator of who the
secret is from compared to just a name. It also future proofs in case
the secret sender is a service rather than an individual.