[#20240626] Replace `format` with basic var interpolation
Not the result of a broader decision about how to put string together.
Right now we're specifically trying to figure out where a malformed
string '%' error is coming from. This is simply to eliminate one
possibility. (Side note: it's looking like the issue is with a dynamic
string that's being passed in, not with the format string itself. And
not necessarily any lines in these changes.)