[#510] Update otto dependency
This update moves the otto gem from version 1.0.2 to the pre-release
version 1.1.0.pre.alpha1. This change may introduce new features or
improvements, but it's important to note that it's an alpha release and
could potentially include breaking changes or unstable functionality.
Testing is recommended to ensure compatibility with the current
codebase.