[#510] Include shrimp in near miss API responses
e.g. When there's a form error, that means the shrimp for that request
was good. If the UI from a Vue component wants to try again it would
need to refresh the entire page to get a fresh shrimp. This change
returns the fresh shrimp with the form error. Successful responses
include the shrimp as well.