[#510] Use OT logging methods
Previously syslog. I love syslog. But when running containerized
services, it's a lot more convenient to log to STDOUT. Otherwise the
host needs to look at the syslog inside the container in addition to the
`docker logs` (or whatever is running the containers).

Here, using the OT methods is obviously specific to this codebase but
then again so is having your source code tracked in the same repo. Long
story short, I don't mind and prefer it over overthinking a properly
general solution.