[#541] Fix for errant  "null"
This didn't raise an exception prior to the familia v1.0 upgrade. The
Metadata class included RedisHash which relied on method_missing to
fetch values from the... well redis hash. We've never had any fields
named "null" so `metadata.null` has always returned `nil`.

The messy abstraction that RedisHash provided was the initial impetus
for revisiting Familia in the first place so it's amusing, rewarding,
and frightening to find this so close to the end of the work. 🤘

(Frightening only in terms of how much worse off we could have been all
this time!)