[#648] Split api routes in v1 & v2
In a previous minor release, we added new base API methods like
`retrieve_records`, `process_action` when implementing new endpoints for
account and domains. At the time we didn't think it was enough of a
change to warrant labelling them as v2. We had "bigger" changes in mind
but realistically it's a new approach to serving data via the API and
we're not going to be updating the pre-existing v1 routes to use them so
v2 is the logical home for them.

These modifications imply that a v3 API would be a consolidated,
all-json API. Consolidated as in, it'd function as both our "internal"
API for our own front-end as well as for our "external" API for
customers to enjoy.