[#713] (WIP) Add v1 definition to v2 path
This is going to take a few janky steps to get from where we are to
where we need to be. The long and the short is that Postman only allows
one git repo connection for Free and Basic plans: "Upgrade your plan to
link this API to a repository". I have no qualms about upgrading to the
Professional plan but that's the only thing we need from it and it's
otherwise twice the price. After a bit of searching, the limitation was
added with Postman v10 (currently v11).

Note that repo Sync isn't mentioned on the pricing page and the
limitation isn't mentioned on either of the docs pages that talk about
why and how to set it up. This feels like a real Elaine Benes moment but
it's the principle of the thing.

https://www.postman.com/pricing/

https://learning.postman.com/docs/designing-and-developing-your-api/versioning-an-api/using-cloud-git-repo/
https://learning.postman.com/docs/designing-and-developing-your-api/versioning-an-api/connecting-multiple-apis/