[#906] Update eslint rule for selfclosing tags
Align with existing codebase which predominantly does use a newline in
this case. NOTE: In Zed, Vue.js files have auto-format on save disabled
which coincides with this change, otherwise prettier will replace the newline
with a space. AFAICT there is currently no way to distinguish prettier
settings for selfclosing tags to properly align both tools.

e.g. Putting `/>` on its own line is a nice way of dignifying selfclosing
tags (and easier to follow when scanning the code). The same is not true
for a `>` regular closing tag which is dangerous as a dangler.

Signed-off-by: delano <delano@onetimesecret.com>