[#878] Parse server injected data immediately
It seems to behave more reliably when running prior to any of the Vue
code. I've moved it to main.ts preior to creating and initializing the
app and can still fail with the error thrown from WindowService.get. I'd
considered AppInitializer but it's too late since all the code in
App.vue would run before (it runs when the app is created
`createApp(App)`). We would have to be very specific about not calling
on the window service.

With this approach, the data is just immediately available at
window.__ONETIME_STATE__ when Vue starts.