[#1018] Properly parse brand schema in identity store
Without parsing the schema, the types were all loose strings since these
settings come via the window service rather than the API. We do that so
we can instantly show the correct page details without the delay of a
request. The homepage toggle is working correctly. The branded homepage
wasn't toggling the secret link form on and off b/c `"true"` isn't
`true` as we expect when dealing with data via the API.