[#1142] Add locale support to auth forms
* Adds locale prop to SignIn and SignUp form components
* Propagates current locale from language store to auth forms
* Includes hidden locale input field in form submissions
* Sets default locale to 'en' in both components

Without this, verification emails would still be sent in the default locale.

NOTE: There is a nuance where if the language is changed while on the
form page, it's important that the component be using a reactive
reference to the current locale. Otherwise it will be whatever the
language was before the user changed it.