[#1155] Add text versions of the email templates
We send emails with both HTML and text to ensure they're readable in the
most circumstances. Previously we used the HTML content with the tags
stripped which isn't ideal. This way we can actually control how the
text emails are formatted.

Includes a readme with the templates listed, along with their template
vars for reference.