[#1185] Promote lib/apps dir to project root
This is a strong convention across ruby frameworks. I'd hessitated to
make the change previously but two things encouraged it this time
around: removing the Onetime::App namespace so the resulting paths now
are descriptive not redudant (e.g. like `./apps/api/v1` instead of
`./apps/onetime/app/api/v1/routes` etc); having realized the parity of
the Vue app in ./src and the ruby app(s) in ./apps.