[#1283] Prioritize domain branding locale
* Ensures locale from domain branding takes
  precedence over the current locale.
* Allows specific branding to override the
  default localization settings.

This is a fix for a bug introduced in v0.20.0 related to the view helper
and serializer classes. Previously the base view would set `locale` to
the brand locale if it had a value. With the helper and serializer
registry, they can specify dependencies, but serializers do not have
access to the output from other serializers. IOW, the i18n serializer
would need to reload the custom domain record to get the brand locale.
Instead of doing that, we hook into the frontend locale logic and
prioritize the domain brand "just in time".

There may be other implications with this inadvertent change but perhaps
letting the frontend decide is the Right Way to go anyway (rather than
setting locale on the backend and the frontend just loading whatever
it's told to).