[#1380] Clean up intentionally wack retroactive testcases
The tests that we added to v0.20.5 (while working on v0.22.1) were in
some cases written to workaround the boot-time configuration bugs that
existed prior to the v0.21.0 refactor and that we were working on fixing
in v0.22.1. The workarounds were necessary in order to write the tests
(e.g. mimicing the complexity of OT.boot! using the input config and
OT.conf interchangably (wild). These workarounds are no longer necessary
and the tests can be simplified. This is a cleanup of those tests to
remove the workarounds and make them more readable.

Note: this is a wip. More cleanup to come.