[#1452] Clean up file naming
Rename `onetime/initializers/boot.rb` to `onetime/boot.rb` and
`onetime/classmethods.rb` to `onetime/class_methods.rb`.

Update `onetime.rb` to require the new files and move the `extend` calls
into their respective files.

Reasoning to move boot out of initializers is that it is called
externally (OT.boot!) and is aware of the initializers as a whole. But
still separate from class_methods.rb bc it's so prominent and should
standout in the file system as an obvious place to look for start up
behaviour.