[#1452] Refine migration dry run output
Adjust the CLI output messages to clearly indicate if a migration
was a dry run. Also, change the return value of Onetime::Migration.run
to false when a migration is not needed and it's a dry run. This ensures
the exit code and output correctly signal that no migration was run.