[#1481] Rename OT::Config to OT::Configurator
Renamed the Onetime::Config class to Onetime::Configurator for better overall clarity. I was considering ConfigProcessor as well which is a better name on its own but is actually more confusing elsewhere in the codebase.  For example, we want to keep a reference to the instantiated class at boot time:

OT.conf is the regular configuration hash with all of the actual settings. So that name is taken.

OT.processor is awful since it can mean different things.
OT.config_processor is a lot better, but it's long and still associated with an overloaded term "processor".
OT.configurator is honestly not amazing on its own. but it is distinctive. It's safe to assume that it's an instance of the OT::Configurator class. Even if you don't know what a Configurator is exactly, you know it's to do with configuration and you know where to look next for more information.