[#1481] Improve JSON support to readiness middleware
Can now be subclassed and have the response type pre-set, rather than relying on request headers. For example a browser request for an API endpoint won't mention accepting application/json content. The only way for the API endpoint to know what format to respond in is how the middleware was initialized.