[#1461] Eliminate support host setting
The settings has been... eliminated.

It was a stop-gap measure for linking to a documentation site in a way
that wasn't specific to onetimesecret.com. Since the footer links are
now configurable (along with similar new functionality), support_host is
no longer needed. And that is why it has been fully... eliminated.