[#1500] "Now let me be clear"
Clarify boot event order and error handling: we pass the frozen static
config to the providers so they can start up and get healthy. We now
only make OT.conf available if they all do that successfully. This
guarantees that if the application is to run in a degraded state, it
really cannot do much of anything. But it definitely can't quietly do
things incorrectly/indeterminately either.

For the moment it also very helpfully forces us to revisit all code on
the boot path -> with OT globals not available, the only way to get
configured up all nice is to accept the ServiceProvider.start(config)
offering.