[#1500] Modernize SystemSettings model
It was one of the earliest areas of code, predating the new boot
and initialization jive. It was originally doing a lot more b/c the
application was using just the single config file and we didn't want
to display the entire config in the Colonel UI. So it was the first
kernel of dynamic config functionality.

Anyway, long story short these changes clean up some of that and
correct field naming. I've just commented out the to_onetime_config
method for now in the unlikely event that we'll need it again. It
outputs the dynamic config in a structure that matches the old YAML.