[#1508] Remove config markdown docs
They're a lot to maintain and the config is still changing so much.
It's actually more confusing with inconsistencies than nothing at all
and too easy for them to get outdated. Left the CONFIGURATION.md doc
linked to from the main readme since it's focus is on installation and
running (which are practical documents -- not comprehensive sources of truth).