[#1512] Fix user_type data migration
We manually set obj.user_type to authenticated so "undo" the default
behaviour of setting the field to anonymous in the model init. That's
the right thing to do in the init (and is the same behaviour we used
with custid=anon) but in this specific case for the initial data
migration, the obj.user_type is returning 'anonymous' since the
field doesn't exist yet.

We also remove the cust.anonymous? filtering. It's unnecessary since
we know that there should be no records in the DB where custid=anon.
If there are then we want to remove those anyway.