[#1523] Excoriate indifferent hash access
This simplifies hash access and provides behaviour that is predictable
at the code of a few quotation marks.

I like indifferent hash access in theory. But even as a refinement
it's still too fussy which makes any implementing code delicate in
unnecessary ways. Like complicating freezing a hash at start time. And
particularly for hashes of hashes. We had a situation where accessing
`conf[:site][:secret]` with any permutation of symbol or string
arguments always returned nil even though `conf[:site].fetch('secret')`
returned `false`. It was due to a bug in `def []` for any value that
was falsy so it wasn't the fault of indifferent hash access as a
concept. It just opened the door for a really silly, awful, hard to
debug behaviour to waltz in. I can deal with a few quotation marks.