[#1523] Move refinement to v1 home
This code although deprecated is still kept for v1 compatibility. The
legacy v1 API is in maintenance mode warts and all. It's been around for
over 10 years so there are a lot of inbound integrations with other systems.
Until we have a specific migration plan in place, this module will continue
to be maintained for the society of creative anachronisms.