Replace direct OT.global_secret access
The new config system has done away with a bazillion properties on
the Onetime module. The value is still available via the config in the
same place as always site->secret. We may want to set the site_secret
to the shared state though which would make it easier to test (since
all config is immutable now).