[#1554] Remove obsolete change_email command
Soon to be obsolete. It won't be the rigamarole it is now b/c both
domains and customers will use generated object IDs so changing the
email address is just a matter of updating a field. It'll be self-
serve via new auth system as well (as it should be! we're not telling
any tales out of school here).