[#1552] [#1555] Use lowercase HTTP headers in Vue app for consistency
Ideally we'll be able to use lowercase headers in the frontend as
well b/c that would simplify things across codebases. I've gone
through and made the relevant changes in the context of where we
need them, but in a separate commit that we can revert if need be.

Related to previous commit 385d3e2a3a846db2aa731dc15141dd875eb3e2ba