[#1557] Fix serialized hashes in MutableConfig.create
It was trying to save each config section as the ruby hash, I guess
assuming that Familia would handle the JSON serialization. It does
usually but not when using the base metal command methods like hset.
Depending on how we get the mutable config sections, we may need to
de-serialize as well.

Although it would all be much easier if we updated Familia directly
to support multi-saves by default. That would allow us to use the
regular, familiar horreum methods and impress everyone in computer
class.