[#1554] Fix for familia horreum field bug
NOTE: Due to a limitation in Familia::Horreum (when a field is defined
its read accessor is created the same way regardless of whether there
was an existing accessor), the model definitions must live at the top.
More specifically, the horreum field definitions must be loaded prior
to the model instance methods.