[#1554] Use fetch instead of []
The FlexibleHashAccess refinement overrides fetch and dig methods to
handle string/symbol "indifferency" (a new Claude-ism), but it doesn't
override [].

This is a cheesy fix. I'd rather remove the refinement altogether like
we did for v2 API and lib code. The thinking was to leave v1 well
enough alone to avoid adding bugs to sunset code. That also leaves
the door open for situations like this where we "dig" in to some
quirky behaviour only to find out, "Oh we made X change over in Y and
ended up breaking Z anyway".

We'll get the test suites passing to merge this branch work in so we
can get to the Familia and Rack upgrades using the existing passing
tests as our guide. Then remove FlexibleHashAccess at that point.